Fix test12 and one of testaa's issues. Enable tests that pass on 64 bit now.

+++ test/runnable/testaa.d
@@ -193,7 +193,7 @@ struct Key
 
     static int hash, cmp, equals;
 
-    uint toHash() const
+    size_t toHash() const
     {	hash = 1;
 	return a + b + c + d;
     }
